Animate top property instead of transform and margin
Using margin to offset the modal relative to the page uses the document
width instead of the height, so a user could have a screen that is small
enough that they wouldn't be able to see the modal. Top and bottom
properties using percentage values are based on the height of the
document, so using `top` instead of `margin` will guarantee us that the
modal will *always* be visible regardless of the user viewport height.